requirements.txt
python-fasthtml
requests
